chore: move QuestType to enums folder
