+++ src/database/enums/QuestType.enum.ts
@@ -1,0 +1,12 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum QuestType {
+  GENESIS = 'GENESIS',
+  PRIME = 'PRIME',
+  DAILY = 'DAILY',
+  PROGRESSIVE = 'PROGRESSIVE',
+}
+
+registerEnumType(QuestType, {
+  name: 'QuestType',
+});
